functions/requirements.txt
firebase_functions~=0.1.0
openai
python-dotenv
google-cloud-logging